Rename also module
