+++ src/videomeg_browser/syncable_browser.py
